Don't close preferences when cancelling tour
When the user clicks the XKit button, they're either curious what XKit can do, in which case they'll probably either take the tour or explore on their own; or they want to get at the preferences. Closing out the preferences panel if they cancel the tour is going to be bad UX in any of those cases.